feat(sapphirecorev1.sol): implement deposit function
Passes the original deposit functions. Added the `vaults` mapping to the storage and
`collateralLimit`. Changed `ActionOperated` to `ActionsOperated`. Added the collateral limit in the
`setLimits()` function. Changed the Vault struct to contain `uint256` instead of Amount.Principal.
Added the collateral limit to the sapphire defaults. Now calls `core.setLimits` in
`setupSapphire()`.