refactor(calculations.ts): fix import error
BREAKING CHANGE: This commit does not include the breaking change per se, but this branch does. It
includes all the refactoring for the added `protocol` in the ScoreProof struct. This is a breaking
change because now all files that previously used a credit score now have to check for the protocol.
In addition, all these files must store a "protocol" value to check against.